librecover: compile as shared library rather than static
Linking the library as a shared object initialises the C++
standard objects (e.g. `std::cout`) even if it is linked
with C code.

This fixes a bug where multiple C++ objects like `std::cout`
would segfault because they wouldn't be initialised. The
initialisation code is added by the C++ compiler when
compiling the `main` function [1]. Since the rest of uftrace
is written in C and compiled by a C compiler, the
initialisation code wouldn't be added when linking
the library as static.

[1] https://isocpp.org/wiki/faq/mixing-c-and-cpp#c-calls-cpp