mcount: dynamic: Reset patterns on each do_dynamic_update
The dynamic patching/unpatching of a function is done by
calling do_dynamic_update each time. This function appends
the filter pattern to a global list and then filter all
functions against all patterns. After multiple patching
and unpatching, this list can becomes large.

While it may not necessarly be a performance problems,
there is no need to keep track of old patterns each time
we patch/unpatch. This commit resets the list of patterns.

Finally, two hashmaps are added to keep track of which
functions are patched or unpatched.